Refactor Dtos
