+++ backend/src/main/java/com/hot6/pnureminder/controller/MainTestController.java
@@ -1,18 +1,13 @@
 package com.hot6.pnureminder.controller;
 
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.io.IOException;
-
 @RestController
 @RequestMapping("/api")
-public class HelloController {
+public class MainTestController {
 
     @GetMapping("/main")
     public ResponseEntity<String> tester(){
