+++ backend/src/main/java/com/hot6/pnureminder/dto/Token/TokenReqDto.java
@@ -1,10 +1,10 @@
-package com.hot6.pnureminder.dto;
+package com.hot6.pnureminder.dto.Token;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class TokenRequestDto {
+public class TokenReqDto {
     private String accessToken;
     private String refreshToken;
 }
